HYD-4292 Create dirp; a source loader
There are various places in the application and in our tests in which we
load source files from a directory. Specifying these files manually
becomes difficult to maintain. Dirp allows us to specify a directory and
the list of files to load and very quickly be returned an object
containing the contents of each source file. This leads to more
maintainable code and prevents us from having to update multiple files
when adding a new source.

In this patch I am removing the lodash dependency, adding the fp
library and updating the source to use ES 5 and fp to accomplish
the same thing.

  - index.js - Updated source to perform the same operationg but without
  lodash.
  - package.js - Added fp and removed lodash
  - tests/tests.js - renamed dirp-tests.js to tests.js
  - jasmine.json - Updated to point to tests correctly without the -spec
    string.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>

- Took out the pathFromDir variable as we no longer need to curry and
instead put it inline when mapping over filenames.