HYD-4707 Update Dirp such that it can only read .json and .js files.
We noticed an issue where dirp was attempting to load a readme file.
Dirp should only try to load .json and .js files. Furthermore, it
should continue to use a blacklist of .js and .json files that
should not be loaded.

This patch will ensure that only .js and .json files are loaded by
checking the file extension of each file. I will also update the
test to ensure that the files that are not supposed to be loaded
are not included in the results.

  - index.js - Add list of valid file extensions and check the file
  extension of each file being passed in.
  - tests/tests.js - Add section to ensure files that are not
  supposed to be loaded are not included in the results.
  - .gitignore - Exclude .idea and .iml files (Intellij generated
  files)

Signed-off-by: Will Johnson <william.c.johnson@intel.com>